requirements.txt
langgraph
openai
typing-extensions
python-dotenv
streamlit
requests 